Offsets committer log (#1031)
* Message sending result handling logging, queues utilization metric fix
* Added skipped offsets metric log
* Logging when ConsumerProcess signal queue full
* Added message id to logs

+++ hermes-consumers/src/main/java/pl/allegro/tech/hermes/consumers/queue/MonitoredMpscQueue.java
@@ -20,7 +20,7 @@ public class MonitoredMpscQueue<T> {
         this.queue = new MpscArrayQueue<>(capacity);
         this.name = name;
         this.metrics = metrics;
-        metrics.registerGauge("queue." + name + ".utilization", () -> queue.size() / queue.capacity());
+        metrics.registerGauge("queue." + name + ".utilization", () -> (double) queue.size() / queue.capacity());
     }
 
     public boolean offer(T element) {
